temporary_resource_deletion/deletion_function/requirements.txt
functions-framework==3.*
google-auth
google-cloud-storage
google-cloud-tasks